Add Header Files for Spell Classes
