+++ src/lib/importex/obj.cpp
@@ -911,12 +911,12 @@ void importObj( NifModel * nif, const QModelIndex & index )
 
 					if ( ix == points.count() ) {
 						points.append( p );
-						verts.append( overts.value( ix ) );
-						norms.append( onorms.value( ix ) );
-						texco.append( otexco.value( ix ) );
+						verts.append( overts.value( p.v ) );
+						norms.append( onorms.value( p.n ) );
+						texco.append( otexco.value( p.t ) );
 					}
 
-					tri[t] = p.v;
+					tri[t] = ix;
 				}
 
 				triangles.append( tri );
