+++ src/spells/blocks.h
@@ -1,15 +1,13 @@
 #ifndef SP_BLOCKS_H
 #define SP_BLOCKS_H
-
 #include "spellbook.h"
-
 
 // Brief description is deliberately not autolinked to class Spell
 /*! \file blocks.h
- * \brief Block spells header
- *
- * All classes here inherit from the Spell class.
- */
+* \brief Block spells header
+*
+* All classes here inherit from the Spell class.
+*/
 
 typedef enum {
 	// "nifskope"
@@ -20,7 +18,7 @@ typedef enum {
 	MIME_IDX_VER,
 	// "type"
 	MIME_IDX_TYPE
-} CopyPasteMimeTypes;
+}CopyPasteMimeTypes;
 
 //! Copy a branch (a block and its descendents) to the clipboard
 class spCopyBranch final : public Spell
@@ -29,7 +27,6 @@ public:
 	QString name() const override final { return Spell::tr( "Copy Branch" ); }
 	QString page() const override final { return Spell::tr( "Block" ); }
 	QKeySequence hotkey() const override final { return QKeySequence( QKeySequence::Copy ); }
-
 	bool isApplicable( const NifModel * nif, const QModelIndex & index ) override final;
 	QModelIndex cast( NifModel * nif, const QModelIndex & index ) override final;
 };
@@ -42,11 +39,8 @@ public:
 	QString page() const override final { return Spell::tr( "Block" ); }
 	// Doesn't work unless the menu entry is unique
 	QKeySequence hotkey() const override final { return QKeySequence( QKeySequence::Paste ); }
-
 	QString acceptFormat( const QString & format, const NifModel * nif );
-
 	bool isApplicable( const NifModel * nif, const QModelIndex & index ) override final;
-
 	QModelIndex cast( NifModel * nif, const QModelIndex & index ) override final;
 };
 
@@ -56,9 +50,7 @@ public:
 	QString name() const override final { return Spell::tr( "Duplicate Branch" ); }
 	QString page() const override final { return Spell::tr( "Block" ); }
 	QKeySequence hotkey() const override final { return{ Qt::CTRL + Qt::Key_D }; }
-
 	bool isApplicable( const NifModel * nif, const QModelIndex & index ) override final;
-
 	QModelIndex cast( NifModel * nif, const QModelIndex & index ) override final;
 };
 
@@ -69,11 +61,9 @@ public:
 	QString name() const override final { return Spell::tr( "Remove Branch" ); }
 	QString page() const override final { return Spell::tr( "Block" ); }
 	QKeySequence hotkey() const override final { return{ Qt::CTRL + Qt::Key_Delete }; }
-
 	bool isApplicable( const NifModel * nif, const QModelIndex & index ) override final;
 	QModelIndex cast( NifModel * nif, const QModelIndex & index ) override final;
 };
-
 
 //! Add a link to the specified block to a link array
 /*!
@@ -95,10 +85,10 @@ void delLink( NifModel * nif, const QModelIndex & iParent, QString array, int li
 
 //! Link one block to another
 /*!
- * @param nif The model
- * @param index The block to link to (becomes parent)
- * @param iBlock The block to link (becomes child)
- */
+* @param nif The model
+* @param index The block to link to (becomes parent)
+* @param iBlock The block to link (becomes child)
+*/
 void blockLink( NifModel * nif, const QModelIndex & index, const QModelIndex & iBlock );
 
 #endif // SP_BLOCKS_H
