+++ src/spells/moppcode.cpp
@@ -1,99 +1,49 @@
-#include "spellbook.h"
+#include "moppcode.h"
 
-
-// Brief description is deliberately not autolinked to class Spell
-/*! \file moppcode.cpp
- * \brief Havok MOPP spells
- *
- * Note that this code only works on the Windows platform due an external
- * dependency on the Havok SDK, with which NifMopp.dll is compiled.
- *
- * Most classes here inherit from the Spell class.
- */
-
-// Need to include headers before testing this
-#ifdef Q_OS_WIN32
-
-// This code is only intended to be run with Win32 platform.
-
-extern "C" void * __stdcall SetDllDirectoryA( const char * lpPathName );
-extern "C" void * __stdcall LoadLibraryA( const char * lpModuleName );
-extern "C" void * __stdcall GetProcAddress ( void * hModule, const char * lpProcName );
-extern "C" void __stdcall FreeLibrary( void * lpModule );
-
-//! Interface to the external MOPP library
-class HavokMoppCode
+bool HavokMoppCode::Initialize()
 {
-private:
-	typedef int (__stdcall * fnGenerateMoppCode)( int nVerts, Vector3 const * verts, int nTris, Triangle const * tris );
-	typedef int (__stdcall * fnGenerateMoppCodeWithSubshapes)( int nShapes, int const * shapes, int nVerts, Vector3 const * verts, int nTris, Triangle const * tris );
-	typedef int (__stdcall * fnRetrieveMoppCode)( int nBuffer, char * buffer );
-	typedef int (__stdcall * fnRetrieveMoppScale)( float * value );
-	typedef int (__stdcall * fnRetrieveMoppOrigin)( Vector3 * value );
-
-	void * hMoppLib;
-	fnGenerateMoppCode GenerateMoppCode;
-	fnRetrieveMoppCode RetrieveMoppCode;
-	fnRetrieveMoppScale RetrieveMoppScale;
-	fnRetrieveMoppOrigin RetrieveMoppOrigin;
-	fnGenerateMoppCodeWithSubshapes GenerateMoppCodeWithSubshapes;
-
-public:
-	HavokMoppCode() : hMoppLib( 0 ), GenerateMoppCode( 0 ), RetrieveMoppCode( 0 ), RetrieveMoppScale( 0 ),
-		  RetrieveMoppOrigin( 0 ), GenerateMoppCodeWithSubshapes( 0 )
-	{
+	if ( !hMoppLib ) {
+		SetDllDirectoryA( QCoreApplication::applicationDirPath().toLocal8Bit().constData() );
+		hMoppLib = LoadLibraryA( "NifMopp.dll" );
+		GenerateMoppCode   = (fnGenerateMoppCode)GetProcAddress( hMoppLib, "GenerateMoppCode" );
+		RetrieveMoppCode   = (fnRetrieveMoppCode)GetProcAddress( hMoppLib, "RetrieveMoppCode" );
+		RetrieveMoppScale  = (fnRetrieveMoppScale)GetProcAddress( hMoppLib, "RetrieveMoppScale" );
+		RetrieveMoppOrigin = (fnRetrieveMoppOrigin)GetProcAddress( hMoppLib, "RetrieveMoppOrigin" );
+		GenerateMoppCodeWithSubshapes = (fnGenerateMoppCodeWithSubshapes)GetProcAddress( hMoppLib, "GenerateMoppCodeWithSubshapes" );
 	}
 
-	~HavokMoppCode()
-	{
-		if ( hMoppLib )
-			FreeLibrary( hMoppLib );
-	}
+	return (GenerateMoppCode && RetrieveMoppCode && RetrieveMoppScale && RetrieveMoppOrigin);
+}
 
-	bool Initialize()
-	{
-		if ( !hMoppLib ) {
-			SetDllDirectoryA( QCoreApplication::applicationDirPath().toLocal8Bit().constData() );
-			hMoppLib = LoadLibraryA( "NifMopp.dll" );
-			GenerateMoppCode   = (fnGenerateMoppCode)GetProcAddress( hMoppLib, "GenerateMoppCode" );
-			RetrieveMoppCode   = (fnRetrieveMoppCode)GetProcAddress( hMoppLib, "RetrieveMoppCode" );
-			RetrieveMoppScale  = (fnRetrieveMoppScale)GetProcAddress( hMoppLib, "RetrieveMoppScale" );
-			RetrieveMoppOrigin = (fnRetrieveMoppOrigin)GetProcAddress( hMoppLib, "RetrieveMoppOrigin" );
-			GenerateMoppCodeWithSubshapes = (fnGenerateMoppCodeWithSubshapes)GetProcAddress( hMoppLib, "GenerateMoppCodeWithSubshapes" );
-		}
+QByteArray HavokMoppCode::CalculateMoppCode( QVector<Vector3> const & verts, QVector<Triangle> const & tris, Vector3 * origin, float * scale )
+{
+	QByteArray code;
 
-		return (GenerateMoppCode && RetrieveMoppCode && RetrieveMoppScale && RetrieveMoppOrigin);
-	}
+	if ( Initialize() ) {
+		int len = GenerateMoppCode( verts.size(), &verts[0], tris.size(), &tris[0] );
 
-	QByteArray CalculateMoppCode( QVector<Vector3> const & verts, QVector<Triangle> const & tris, Vector3 * origin, float * scale )
-	{
-		QByteArray code;
+		if ( len > 0 ) {
+			code.resize( len );
 
-		if ( Initialize() ) {
-			int len = GenerateMoppCode( verts.size(), &verts[0], tris.size(), &tris[0] );
+			if ( 0 != RetrieveMoppCode( len, code.data() ) ) {
+				if ( scale )
+				RetrieveMoppScale( scale );
 
-			if ( len > 0 ) {
-				code.resize( len );
-
-				if ( 0 != RetrieveMoppCode( len, code.data() ) ) {
-					if ( scale )
-						RetrieveMoppScale( scale );
-
-					if ( origin )
-						RetrieveMoppOrigin( origin );
-				} else {
-					code.clear();
-				}
+				if ( origin )
+				RetrieveMoppOrigin( origin );
+			} else {
+				code.clear();
 			}
 		}
-
-		return code;
 	}
 
-	QByteArray CalculateMoppCode( QVector<int> const & subShapesVerts,
-	                              QVector<Vector3> const & verts,
-	                              QVector<Triangle> const & tris,
-	                              Vector3 * origin, float * scale )
+	return code;
+}
+
+QByteArray HavokMoppCode::CalculateMoppCode( QVector<int> const & subShapesVerts,
+	QVector<Vector3> const & verts,
+	QVector<Triangle> const & tris,
+	Vector3 * origin, float * scale )
 	{
 		QByteArray code;
 
@@ -101,19 +51,19 @@ public:
 			int len;
 
 			if ( GenerateMoppCodeWithSubshapes )
-				len = GenerateMoppCodeWithSubshapes( subShapesVerts.size(), &subShapesVerts[0], verts.size(), &verts[0], tris.size(), &tris[0] );
+			len = GenerateMoppCodeWithSubshapes( subShapesVerts.size(), &subShapesVerts[0], verts.size(), &verts[0], tris.size(), &tris[0] );
 			else
-				len = GenerateMoppCode( verts.size(), &verts[0], tris.size(), &tris[0] );
+			len = GenerateMoppCode( verts.size(), &verts[0], tris.size(), &tris[0] );
 
 			if ( len > 0 ) {
 				code.resize( len );
 
 				if ( 0 != RetrieveMoppCode( len, code.data() ) ) {
 					if ( scale )
-						RetrieveMoppScale( scale );
+					RetrieveMoppScale( scale );
 
 					if ( origin )
-						RetrieveMoppOrigin( origin );
+					RetrieveMoppOrigin( origin );
 				} else {
 					code.clear();
 				}
@@ -122,34 +72,27 @@ public:
 
 		return code;
 	}
-}
-TheHavokCode;
 
-//! Update Havok MOPP for a given shape
-class spMoppCode final : public Spell
-{
-public:
-	QString name() const override final { return Spell::tr( "Update MOPP Code" ); }
-	QString page() const override final { return Spell::tr( "Havok" ); }
 
-	bool isApplicable( const NifModel * nif, const QModelIndex & index ) override final
+
+	bool spMoppCode::isApplicable( const NifModel * nif, const QModelIndex & index )
 	{
 		if ( nif->getUserVersion() != 10 && nif->getUserVersion() != 11 )
-			return false;
+		return false;
 
 		if ( TheHavokCode.Initialize() ) {
 			//QModelIndex iData = nif->getBlock( nif->getLink( index, "Data" ) );
 
 			if ( nif->isNiBlock( index, "bhkMoppBvTreeShape" ) ) {
 				return ( nif->checkVersion( 0x14000004, 0x14000005 )
-				         || nif->checkVersion( 0x14020007, 0x14020007 ) );
+				|| nif->checkVersion( 0x14020007, 0x14020007 ) );
 			}
 		}
 
 		return false;
 	}
 
-	QModelIndex cast( NifModel * nif, const QModelIndex & iBlock ) override final
+	QModelIndex spMoppCode::cast( NifModel * nif, const QModelIndex & iBlock )
 	{
 		if ( !TheHavokCode.Initialize() ) {
 			Message::critical( nullptr, Spell::tr( "Unable to locate NifMopp.dll" ) );
@@ -168,7 +111,7 @@ public:
 		QModelIndex ihkPackedNiTriStripsData = nif->getBlock( nif->getLink( ibhkPackedNiTriStripsShape, "Data" ) );
 
 		if ( !nif->isNiBlock( ihkPackedNiTriStripsData, "hkPackedNiTriStripsData" ) )
-			return iBlock;
+		return iBlock;
 
 		QVector<int> subshapeVerts;
 
@@ -203,84 +146,73 @@ public:
 
 		if ( verts.isEmpty() || triangles.isEmpty() ) {
 			Message::critical( nullptr, Spell::tr( "Insufficient data to calculate MOPP code" ),
-				Spell::tr("Vertices: %1, Triangles: %2").arg( !verts.isEmpty() ).arg( !triangles.isEmpty() )
-			);
-			return iBlock;
-		}
-
-		Vector3 origin;
-		float scale;
-		QByteArray moppcode = TheHavokCode.CalculateMoppCode( subshapeVerts, verts, triangles, &origin, &scale );
-
-		if ( moppcode.size() == 0 ) {
-			Message::critical( nullptr, Spell::tr( "Failed to generate MOPP code" ) );
-		} else {
-			QModelIndex iCodeOrigin = nif->getIndex( ibhkMoppBvTreeShape, "Origin" );
-			nif->set<Vector3>( iCodeOrigin, origin );
-
-			QModelIndex iCodeScale = nif->getIndex( ibhkMoppBvTreeShape, "Scale" );
-			nif->set<float>( iCodeScale, scale );
-
-			QModelIndex iCodeSize = nif->getIndex( ibhkMoppBvTreeShape, "MOPP Data Size" );
-			QModelIndex iCode = nif->getIndex( ibhkMoppBvTreeShape, "MOPP Data" ).child( 0, 0 );
-
-			if ( iCodeSize.isValid() && iCode.isValid() ) {
-				nif->set<int>( iCodeSize, moppcode.size() );
-				nif->updateArray( iCode );
-				nif->set<QByteArray>( iCode, moppcode );
-			}
-		}
-
+			Spell::tr("Vertices: %1, Triangles: %2").arg( !verts.isEmpty() ).arg( !triangles.isEmpty() )
+		);
 		return iBlock;
 	}
-};
+
+	Vector3 origin;
+	float scale;
+	QByteArray moppcode = TheHavokCode.CalculateMoppCode( subshapeVerts, verts, triangles, &origin, &scale );
+
+	if ( moppcode.size() == 0 ) {
+		Message::critical( nullptr, Spell::tr( "Failed to generate MOPP code" ) );
+	} else {
+		QModelIndex iCodeOrigin = nif->getIndex( ibhkMoppBvTreeShape, "Origin" );
+		nif->set<Vector3>( iCodeOrigin, origin );
+
+		QModelIndex iCodeScale = nif->getIndex( ibhkMoppBvTreeShape, "Scale" );
+		nif->set<float>( iCodeScale, scale );
+
+		QModelIndex iCodeSize = nif->getIndex( ibhkMoppBvTreeShape, "MOPP Data Size" );
+		QModelIndex iCode = nif->getIndex( ibhkMoppBvTreeShape, "MOPP Data" ).child( 0, 0 );
+
+		if ( iCodeSize.isValid() && iCode.isValid() ) {
+			nif->set<int>( iCodeSize, moppcode.size() );
+			nif->updateArray( iCode );
+			nif->set<QByteArray>( iCode, moppcode );
+		}
+	}
+
+	return iBlock;
+}
 
 REGISTER_SPELL( spMoppCode )
 
-//! Update MOPP code on all shapes in this model
-class spAllMoppCodes final : public Spell
+bool spAllMoppCodes::isApplicable( const NifModel * nif, const QModelIndex & idx )
 {
-public:
-	QString name() const override final { return Spell::tr( "Update All MOPP Code" ); }
-	QString page() const override final { return Spell::tr( "Batch" ); }
+	if ( nif && nif->getUserVersion() != 10 && nif->getUserVersion() != 11 )
+	return false;
 
-	bool isApplicable( const NifModel * nif, const QModelIndex & idx ) override final
-	{
-		if ( nif && nif->getUserVersion() != 10 && nif->getUserVersion() != 11 )
-			return false;
-
-		if ( TheHavokCode.Initialize() ) {
-			if ( nif && !idx.isValid() ) {
-				return ( nif->checkVersion( 0x14000004, 0x14000005 )
-				         || nif->checkVersion( 0x14020007, 0x14020007 ) );
-			}
+	if ( TheHavokCode.Initialize() ) {
+		if ( nif && !idx.isValid() ) {
+			return ( nif->checkVersion( 0x14000004, 0x14000005 )
+			|| nif->checkVersion( 0x14020007, 0x14020007 ) );
 		}
-
-		return false;
 	}
 
-	QModelIndex cast( NifModel * nif, const QModelIndex & ) override final
-	{
-		QList<QPersistentModelIndex> indices;
+	return false;
+}
 
-		spMoppCode TSpacer;
+QModelIndex spAllMoppCodes::cast( NifModel * nif, const QModelIndex & )
+{
+	QList<QPersistentModelIndex> indices;
 
-		for ( int n = 0; n < nif->getBlockCount(); n++ ) {
-			QModelIndex idx = nif->getBlock( n );
+	spMoppCode TSpacer;
 
-			if ( TSpacer.isApplicable( nif, idx ) )
-				indices << idx;
-		}
+	for ( int n = 0; n < nif->getBlockCount(); n++ ) {
+		QModelIndex idx = nif->getBlock( n );
 
-		for ( const QModelIndex& idx : indices ) {
-			TSpacer.castIfApplicable( nif, idx );
-		}
-
-		return QModelIndex();
+		if ( TSpacer.isApplicable( nif, idx ) )
+		indices << idx;
 	}
-};
+
+	for ( const QModelIndex& idx : indices ) {
+		TSpacer.castIfApplicable( nif, idx );
+	}
+
+	return QModelIndex();
+}
+
 
 REGISTER_SPELL( spAllMoppCodes )
-
-#endif // Q_OS_WIN32
-
